requirements.txt
streamlit==1.39.0
pandas==2.2.3
nltk==3.9
spacy==3.8.7
python-docx==1.1.2
PyMuPDF==1.24.11
pdfplumber==0.11.2
openpyxl==3.1.5
PyPDF2==3.0.1 
scikit-learn
joblib
numpy